Remove Unneeded Code from the Platformer Demo
Removing unneeded pieces of code - mostly just the bullets and enemies.

I left a few parts in that I think we could reuse - notably, the "Coins" and "Platforms".
While I don't know if we'll use either, I think they _could_ be useful in the future, even if they aren't referenced now.

Also, left the "Shooting" logic in and renamed it as "BlockPlacement", thinking we can reuse that for placing the blocks.

I left all the assets as-is (art, music, sfx) thinking they're good placeholder art until we have everything that we need.

Also, updating the README to explain some of the code.