Change Weather Rates and Make Lava Non-Destructive
Making some pretty fundamental changes - here's my thinking:

I feel like right now, the player doesn't have much incentive to build a structure - it's easy to just hide right before a storm and otherwise just lay blocks down to navigate from here to there.

It's not worth the player's time to build anything because it's too hard to build anything sustainable (the eruptions will tear it apart) and it's easier to just hide when it comes.

So the changes I'm making here are as follows:
* Eruptions are non-descructive (so the player doesn't give up on building)
* Eruptions are much more frequent (5s) but not escalating - so there's a constant threat (and reason to build shelter)
* Make earthquakes just a little more frequent (cause _some_ havoc to what the player builds) and increase the pace of the game

We can tweak these numbers of revert this completely, but I thought this was worth a try